(PC-4937) Fix PR Feedbacks
