+++ src/features/cheatcodes/pages/AppComponents.tsx
@@ -5,8 +5,8 @@ import { TouchableOpacity } from 'react-native-gesture-handler'
 import styled from 'styled-components/native'
 
 import { _ } from 'libs/i18n'
-import AppModal from 'ui/components/modals/AppModal'
-import useModal from 'ui/components/modals/useModal'
+import { AppModal } from 'ui/components/modals/AppModal'
+import { useModal } from 'ui/components/modals/useModal'
 import { ArrowPrevious } from 'ui/icons/ArrowPrevious'
 import { Close } from 'ui/icons/Close'
 import { ColorsEnum, Spacer, Typo } from 'ui/theme'
