+++ src/ui/components/modals/ModalHeader.test.tsx
@@ -4,19 +4,9 @@ import React from 'react'
 import { ArrowPrevious } from 'ui/icons/ArrowPrevious'
 import { Close } from 'ui/icons/Close'
 
-import ModalHeader from '.'
-
-jest.mock('libs/environment', () => ({
-  env: {
-    FEATURE_FLAG_CODE_PUSH: true,
-  },
-}))
+import { ModalHeader } from './ModalHeader'
 
 describe('ModalHeader component', () => {
-  it('should render correctly', () => {
-    const home = render(<ModalHeader title="Testing modal header rendering" />)
-    expect(home).toMatchSnapshot()
-  })
   describe('left icon', () => {
     it('should be hidden when the icon is not provided', () => {
       const { getByTestId } = render(<ModalHeader title="Testing modal header rendering" />)
